backend/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.10


# CORS Headers
django-cors-headers==4.3.1

# Environment Variables
python-decouple==3.8

# Password Hashing
bcrypt==4.1.2

# Timezone Support
pytz==2023.3


